wsd: test: log test assertions
