+++ test/Makefile.am
@@ -101,7 +101,8 @@ unittest_LDADD += -lssl -lcrypto
 unit_base_la_LIBADD += -lssl -lcrypto
 endif
 
-fakesockettest_SOURCES = fakesockettest.cpp ../net/FakeSocket.cpp
+fakesockettest_CPPFLAGS = -g
+fakesockettest_SOURCES = fakesockettest.cpp  ../net/FakeSocket.cpp ../common/Log.cpp ../common/Util.cpp
 fakesockettest_LDADD = $(CPPUNIT_LIBS)
 
 # old-style unit tests - bootstrapped via UnitClient
